+++ src/main/java/org/jsoup/safety/Blacklist.java
@@ -1,0 +1,4 @@
+package org.jsoup.safety;
+
+public class Blacklist {
+}
