+++ src/main/java/org/jsoup/safety/BlacklistCleaner.java
@@ -1,0 +1,37 @@
+package org.jsoup.safety;
+
+import org.jsoup.nodes.Element;
+import org.jsoup.nodes.Node;
+import org.jsoup.select.NodeTraversor;
+import org.jsoup.select.NodeVisitor;
+
+public class BlacklistCleaner extends Cleaner{
+    public BlacklistCleaner(Filter filter) {
+        super(filter);
+    }
+
+    private final class CleaningVisitor implements NodeVisitor {
+        private int numDiscarded = 0;
+        private final Element root;
+        private Element destination; // current element to append nodes to
+
+        private CleaningVisitor(Element root, Element destination) {
+            this.root = root;
+            this.destination = destination;
+        }
+
+        public void head(Node source, int depth) {
+            // TODO
+        }
+
+        public void tail(Node source, int depth) {
+            // TODO
+        }
+    }
+
+    int copySafeNodes(Element source, Element dest) {
+        CleaningVisitor cleaningVisitor = new CleaningVisitor(source, dest);
+        NodeTraversor.traverse(cleaningVisitor, source);
+        return cleaningVisitor.numDiscarded;
+    }
+}
